Move widgets to widgets folder
