+++ lib/views/desktop/send/send.dart
@@ -2,9 +2,9 @@ import 'package:cross_file/cross_file.dart';
 import 'package:dart_wormhole_gui/config/routes/routes.dart';
 import 'package:dart_wormhole_gui/config/theme/colors.dart';
 import 'package:dart_wormhole_gui/constants/app_constants.dart';
-import 'package:dart_wormhole_gui/views/desktop/send/widget/DTCodeGeneration.dart';
-import 'package:dart_wormhole_gui/views/desktop/send/widget/DTSelectAFile.dart';
-import 'package:dart_wormhole_gui/views/desktop/send/widget/DTSendingProgress.dart';
+import 'package:dart_wormhole_gui/views/desktop/send/widgets/DTCodeGeneration.dart';
+import 'package:dart_wormhole_gui/views/desktop/send/widgets/DTSelectAFile.dart';
+import 'package:dart_wormhole_gui/views/desktop/send/widgets/DTSendingProgress.dart';
 import 'package:dart_wormhole_gui/views/desktop/widgets/custom-app-bar.dart';
 import 'package:dart_wormhole_gui/views/shared/send.dart';
 import 'package:dart_wormhole_william/client/native_client.dart';
