+++ lib/views/desktop/send/widgets/DTCodeGeneration.dart
